Minor improvements to loading of nav.dat files
- simgear::strutils::readNonNegativeInt<unsigned int>() is better than
  std::stoul() at detecting and reporting erroneous input (for one, it
  checks that the given string *represents* an integer that fits in the
  specified type, as opposed to just "starts with digits").

- The format version is now an unsigned int (just because it's largely
  enough and simgear::strutils::readNonNegativeInt<unsigned long, 10>()
  isn't instantiated for now).

- The format version is now printed to the log (it was already for
  apt.dat files, but not for nav.dat files).

+++ src/Navaids/navdb.hxx
@@ -58,9 +58,9 @@ class NavLoader {
                                 const std::string& utf8Path,
                                 unsigned int lineNum,
                                 FGPositioned::Type type = FGPositioned::INVALID,
-                                unsigned long version = 810);
+                                unsigned int version = 810);
 };
 
 } // of namespace flightgear
-  
+
 #endif // _FG_NAVDB_HXX
